+++ src/core/flow.test.js
@@ -167,7 +167,7 @@ describe('phases', () => {
     const flow = FlowWithPhases({
       endTurnIf: (G, ctx) => ctx.phase === 'C',
       startingPhase: 'A',
-      phases: { A: {}, B: {}, C: {} },
+      phases: { A: { next: 'B' }, B: { next: 'C' }, C: {} },
     });
 
     let state = { G: {}, ctx: flow.ctx(2) };
@@ -175,12 +175,12 @@ describe('phases', () => {
     expect(state.ctx.phase).toBe('A');
     expect(state.ctx.currentPlayer).toBe('0');
 
-    state = flow.processGameEvent(state, gameEvent('endPhase', 'B'));
+    state = flow.processGameEvent(state, gameEvent('endPhase'));
 
     expect(state.ctx.phase).toBe('B');
     expect(state.ctx.currentPlayer).toBe('0');
 
-    state = flow.processGameEvent(state, gameEvent('endPhase', 'C'));
+    state = flow.processGameEvent(state, gameEvent('endPhase'));
 
     expect(state.ctx.phase).toBe('C');
     expect(state.ctx.currentPlayer).toBe('1');
@@ -214,7 +214,7 @@ test('movesPerTurn', () => {
     state = flow.processMove(state, makeMove('move', null, '0').payload);
     expect(state.ctx.turn).toBe(1);
 
-    state = flow.processGameEvent(state, gameEvent('endPhase', 'B'));
+    state = flow.processGameEvent(state, gameEvent('endPhase', { next: 'B' }));
 
     expect(state.ctx.turn).toBe(1);
     state = flow.processMove(state, makeMove('move', null, '1').payload);
@@ -241,7 +241,7 @@ test('onTurnBegin', () => {
     expect(onTurnBegin).toHaveBeenCalled();
     expect(onTurnBeginOverride).not.toHaveBeenCalled();
 
-    state = flow.processGameEvent(state, gameEvent('endPhase', 'B'));
+    state = flow.processGameEvent(state, gameEvent('endPhase', { next: 'B' }));
     expect(state.ctx.phase).toBe('B');
     expect(onTurnBeginOverride).not.toHaveBeenCalled();
 
@@ -262,7 +262,7 @@ test('onTurnEnd', () => {
     const flow = FlowWithPhases({
       onTurnEnd,
       startingPhase: 'A',
-      phases: { A: {}, B: { onTurnEnd: onTurnEndOverride } },
+      phases: { A: { next: 'B' }, B: { onTurnEnd: onTurnEndOverride } },
     });
 
     let state = { ctx: flow.ctx(2) };
@@ -280,7 +280,7 @@ test('onTurnEnd', () => {
     onTurnEnd.mockReset();
     onTurnEndOverride.mockReset();
 
-    state = flow.processGameEvent(state, gameEvent('endPhase', 'B'));
+    state = flow.processGameEvent(state, gameEvent('endPhase'));
 
     flow.processGameEvent(state, gameEvent('endTurn'));
     expect(onTurnEnd).not.toHaveBeenCalled();
@@ -306,7 +306,7 @@ test('onMove', () => {
     let state = { G: {}, ctx: flow.ctx(2) };
     state = flow.processMove(state, makeMove().payload);
     expect(state.G).toEqual({ A: true });
-    state = flow.processGameEvent(state, gameEvent('endPhase', 'B'));
+    state = flow.processGameEvent(state, gameEvent('endPhase', { next: 'B' }));
     state = flow.processMove(state, makeMove().payload);
     expect(state.G).toEqual({ B: true });
   }
@@ -407,8 +407,8 @@ test('endGameIf', () => {
   }
 });
 
-test('endTurnIf', () => {
-  {
+describe('endTurnIf', () => {
+  test('global', () => {
     const flow = FlowWithPhases({ endTurnIf: G => G.endTurn });
     const game = Game({
       moves: {
@@ -425,9 +425,9 @@ test('endTurnIf', () => {
     expect(state.ctx.currentPlayer).toBe('0');
     state = reducer(state, makeMove('A'));
     expect(state.ctx.currentPlayer).toBe('1');
-  }
+  });
 
-  {
+  test('phase specific', () => {
     const flow = FlowWithPhases({
       startingPhase: 'A',
       phases: {
@@ -449,7 +449,23 @@ test('endTurnIf', () => {
     expect(state.ctx.currentPlayer).toBe('0');
     state = reducer(state, makeMove('A'));
     expect(state.ctx.currentPlayer).toBe('1');
-  }
+  });
+
+  test('return value', () => {
+    const flow = FlowWithPhases({ endTurnIf: () => ({ next: '2' }) });
+    const game = Game({
+      moves: {
+        A: G => G,
+      },
+      flow,
+    });
+    const reducer = CreateGameReducer({ game, numPlayers: 3 });
+
+    let state = reducer(undefined, { type: 'init' });
+    expect(state.ctx.currentPlayer).toBe('0');
+    state = reducer(state, makeMove('A'));
+    expect(state.ctx.currentPlayer).toBe('2');
+  });
 });
 
 test('canMakeMove', () => {
@@ -493,7 +509,7 @@ test('canMakeMove', () => {
   expect(state.G).not.toMatchObject({ C: true });
 
   // Phase B (B is allowed).
-  state = reducer(state, gameEvent('endPhase', 'B'));
+  state = reducer(state, gameEvent('endPhase', { next: 'B' }));
   state.G = {};
   expect(state.ctx.phase).toBe('B');
 
@@ -505,7 +521,7 @@ test('canMakeMove', () => {
   expect(state.G).not.toMatchObject({ C: true });
 
   // Phase C (A and B allowed).
-  state = reducer(state, gameEvent('endPhase', 'C'));
+  state = reducer(state, gameEvent('endPhase', { next: 'C' }));
   state.G = {};
   expect(state.ctx.phase).toBe('C');
 
@@ -517,7 +533,7 @@ test('canMakeMove', () => {
   expect(state.G).not.toMatchObject({ C: true });
 
   // Phase D (A, B and C allowed).
-  state = reducer(state, gameEvent('endPhase', 'D'));
+  state = reducer(state, gameEvent('endPhase', { next: 'D' }));
   state.G = {};
   expect(state.ctx.phase).toBe('D');
 
@@ -596,7 +612,7 @@ test('endGame', () => {
   }
 });
 
-test('endTurn / endPhase args', () => {
+describe('endTurn / endPhase args', () => {
   const flow = FlowWithPhases({
     startingPhase: 'A',
     phases: { A: { next: 'B' }, B: {}, C: {} },
@@ -604,30 +620,41 @@ test('endTurn / endPhase args', () => {
 
   const state = { ctx: flow.ctx(3) };
 
-  {
+  beforeEach(() => {
+    jest.resetAllMocks();
+  });
+
+  test('no args', () => {
     let t = state;
     t = flow.processGameEvent(t, gameEvent('endPhase'));
     t = flow.processGameEvent(t, gameEvent('endTurn'));
     expect(t.ctx.playOrderPos).toBe(1);
     expect(t.ctx.currentPlayer).toBe('1');
     expect(t.ctx.phase).toBe('B');
-  }
+  });
 
-  {
+  test('invalid arg to endPhase', () => {
     let t = state;
     t = flow.processGameEvent(t, gameEvent('endPhase', 'C'));
+    expect(error).toBeCalledWith(`invalid argument to endPhase: C`);
+    expect(t.ctx.phase).toBe('A');
+  });
+
+  test('invalid arg to endTurn', () => {
+    let t = state;
     t = flow.processGameEvent(t, gameEvent('endTurn', '2'));
+    expect(error).toBeCalledWith(`invalid argument to endTurn: 2`);
+    expect(t.ctx.currentPlayer).toBe('0');
+  });
+
+  test('valid args', () => {
+    let t = state;
+    t = flow.processGameEvent(t, gameEvent('endPhase', { next: 'C' }));
+    t = flow.processGameEvent(t, gameEvent('endTurn', { next: '2' }));
     expect(t.ctx.playOrderPos).toBe(2);
     expect(t.ctx.currentPlayer).toBe('2');
     expect(t.ctx.phase).toBe('C');
-  }
-
-  {
-    let t = state;
-    t = flow.processGameEvent(t, gameEvent('endTurn', '0'));
-    expect(t.ctx.playOrderPos).toBe(0);
-    expect(t.ctx.currentPlayer).toBe('0');
-  }
+  });
 });
 
 test('resetGame', () => {
@@ -706,7 +733,7 @@ test('undo / redo restricted by undoableMoves', () => {
   expect(state.G).toEqual({ C: true });
 
   state.G = {};
-  state = reducer(state, gameEvent('endPhase', 'B'));
+  state = reducer(state, gameEvent('endPhase', { next: 'B' }));
   state = reducer(state, gameEvent('endTurn'));
   expect(state.ctx.phase).toBe('B');
 
@@ -724,7 +751,7 @@ test('undo / redo restricted by undoableMoves', () => {
   expect(state.G).toEqual({ C: true });
 
   state.G = {};
-  state = reducer(state, gameEvent('endPhase', 'C'));
+  state = reducer(state, gameEvent('endPhase', { next: 'C' }));
   state = reducer(state, gameEvent('endTurn'));
   expect(state.ctx.phase).toBe('C');
 
@@ -742,7 +769,7 @@ test('undo / redo restricted by undoableMoves', () => {
   expect(state.G).toEqual({ C: true });
 
   state.G = {};
-  state = reducer(state, gameEvent('endPhase', 'D'));
+  state = reducer(state, gameEvent('endPhase', { next: 'D' }));
   state = reducer(state, gameEvent('endTurn'));
   expect(state.ctx.phase).toBe('D');